refatorando a chamada do env url nos testes

+++ cypress/e2e/api/deleteAnOrder_api.cy.js
@@ -1,4 +1,6 @@
 const accessToken = `Bearer ${Cypress.env('api_access_token')}`
+const API_URL = Cypress.env('API_URL')
+
 describe('Deletar um pedido', () => {
 
     let firstOrderID
@@ -6,35 +8,34 @@ describe('Deletar um pedido', () => {
     beforeEach(() => {
         cy.request({
             method: 'GET',
-            url: 'https://simple-books-api.glitch.me/orders',
+            url: `${API_URL}orders`,
+            body: {},
             headers: { Authorization: accessToken },
-        }).then(response => {
-            const responseBody = response.body;
-            expect(response.status).to.equal(200);
-            expect(responseBody).to.be.an('array').and.to.have.length.greaterThan(0)
-            firstOrderID = responseBody[0].id
         })
+        .then(response => {
+            firstOrderID = response.body[0].id
+          });
     })
 
     it('Deve deletar um pedido com sucesso', () => {
 
-        cy.request({
-            method: 'DELETE',
-            url: `https://simple-books-api.glitch.me/orders/${firstOrderID}`,
-            body: {},
-            headers: { Authorization: accessToken },
-        })
-            .then(response => {
-                const responseBody = response.body
-                expect(response.status).to.equal(204)
-
+        if (firstOrderID) {
+            cy.request({
+              method: 'DELETE',
+              url: `${API_URL}orders/${firstOrderID}`,
+              body: {},
+              headers: { Authorization: accessToken },
+            }).then(response => {
+              const responseBody = response.body;
+              expect(response.status).to.equal(204)
             })
+          }
     })
 
     it('Deve checar se o id do pedido já foi deletado', () => {
         cy.request({
             method: 'DELETE',
-            url: "https://simple-books-api.glitch.me/orders/1",
+            url: `${API_URL}orders/1`,
             body: {},
             failOnStatusCode: false,
             headers: { Authorization: accessToken },
